pyproject.toml
[build-system]
requires = ["setuptools>=67.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-yt-dlp"
authors = [{name = "SubFabula"}]
dynamic = ["version"]
requires-python = ">=3.8"
readme = "README.md"
description = '"A feature-rich command-line audio/video downloader" "...with a simple graphical interface".'
dependencies = [
  "yt-dlp",
  "pywebview",
  "ffmpeg",
]

[project.optional-dependencies]
gui = [
  "qtpy",
  "PySide6>=6.9.0",
]

[project.scripts] # This doesn't work for whatever reason.
auto-yt-dlp = "auto_yt_dlp:run"
auto_yt_dlp = "auto_yt_dlp:run"
